\name{bagelR-package}
\alias{bagelR-package}
\alias{bagelR}
\docType{package}
\title{
\packageTitle{bagelR}
}
\description{
\packageDescription{bagelR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bagelR}
\packageIndices{bagelR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{bagelR}

Maintainer: \packageMaintainer{bagelR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
